% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siawd_package.R
\docType{package}
\name{siawd}
\alias{siawd-package}
\alias{siawd}
\title{siawd: Stress in Action Wearables Database App}
\description{
This package contains the Shiny application for the
Stress in Action Wearables Database (SiA-WD).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HugoGit39/siawd}
  \item Report bugs at \url{https://github.com/HugoGit39/siawd/issues}
}

}
\author{
\strong{Maintainer}: Hugo Klarenberg \email{h.klarenberg@vu.nl}

}
